Created a separate reporters folder which has: - A plain reporter - A color reporter The color reporter is given as default argument to check() in pyta.py. The plain reporter can be passes in as an argument if needed.
