+++ pyta.py
@@ -10,27 +10,25 @@ To run the checker, call the check function on the name of the module to check.
 """
 import importlib.util
 import pylint.lint as lint
-from pylint.reporters import BaseReporter
+from reporters.color_reporter import ColorReporter
 
 import webbrowser
 import os
 from urllib.request import pathname2url
 
-import sys
-from colorama import Fore, Style
-from colorama import init
+
 # Local version of website; will be updated later.
 HELP_URL = 'file:' + pathname2url(os.path.abspath('website/index.html'))
 
 
-def check(module_name):
+def check(module_name, reporter=ColorReporter()):
     """Check a module for errors, printing a report.
 
     The name of the module should be passed in as a string,
     without a file extension (.py).
     """
     spec = importlib.util.find_spec(module_name)
-    reporter = PyTAReporter()
+    #reporter = ColorReporter()
     linter = lint.PyLinter(reporter=reporter)
     linter.load_default_plugins()
     linter.load_plugin_modules(['checkers/forbidden_import_checker',
@@ -50,27 +48,4 @@ def doc(msg_id):
     webbrowser.open(msg_url)
 
 
-class PyTAReporter(BaseReporter):
-    def __init__(self):
-        super().__init__(self)
-        self._messages = []
 
-    def handle_message(self, msg):
-        """Handle a new message triggered on the current file."""
-        self._messages.append(msg)
-
-    def print_message_ids(self):
-        # Sort the messages by their type.
-        self._messages.sort(key=lambda s: s[0])
-        # Check if the OS currently running is Windows
-
-        init(wrap=(sys.platform == "win32"))
-        for msg in self._messages:
-            if msg.msg_id.startswith('E'):
-                # Error codes appear in red
-                # code = '\033[1;31m' + msg.msg_id + '\033[0m:'
-                code = Fore.RED + Style.BRIGHT + msg.msg_id + Style.RESET_ALL
-            else:
-                # code = '\033[1m' + msg.msg_id + '\033[0m:'
-                code = Style.BRIGHT + msg.msg_id + Style.RESET_ALL
-            print(code, '({})\n    {}'.format(msg.symbol, msg.msg))
